Start parsing page info

+++ feedbinclient/src/main/kotlin/com/jocmp/feedbinclient/api/FeedbinSubscription.kt
@@ -4,9 +4,9 @@ import com.squareup.moshi.JsonClass
 
 @JsonClass(generateAdapter = true)
 data class FeedbinSubscription(
-    val id: Int,
+    val id: Long,
     val created_at: String,
-    val feed_id: Int,
+    val feed_id: Long,
     val title: String,
     val feed_url: String,
     val site_url: String
